COV^2
The coefficient of variation represents the ratio of the standard deviation to the mean, and it is a useful statistic for comparing the degree of variation from one data series to another, even if the means are drastically different from one another.
Here 'y' columnn is the target column.  Smooth,erratic,lumpy,intermittent these all are the characterics of how the data is , in the SKU.  Its derived from the plot between demand with respect to time v/s variability.